Refactor: types 파일 분할#51
