+++ types/CabinetType.ts
@@ -1,0 +1,10 @@
+export interface BuildingData {
+  building: string;
+  floor: number[];
+  section: string;
+}
+
+export interface SelectedCabinet {
+  cabinetId: number;
+  cabinetNumber: number;
+}
