+++ types/UserType.ts
@@ -1,14 +1,3 @@
-export interface BuildingData {
-  building: string;
-  floor: number[];
-  section: string;
-}
-
-export interface SelectedCabinet {
-  cabinetId: number;
-  cabinetNumber: number;
-}
-
 export interface UserInfo {
   name: string | null;
   affiliation: string | null;
